# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/open-apis/core/enterprise/applist": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "小程序获取用户单位列表(UserAccessToken)",
                "responses": {}
            }
        },
        "/open-apis/core/enterprise/corporate": {
            "put": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "修改单位负责人(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enterprise.PutCorporateReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/open-apis/core/enterprise/create": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "小程序老用户新增安保单位(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enterprise.PostSecurityReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/open-apis/core/enterprise/list": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "获取用户单位列表(UserAccessToken)",
                "responses": {}
            }
        },
        "/open-apis/core/enterprise/new": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "小程序新用户新增安保单位接口",
                "parameters": [
                    {
                        "description": "请求参数体",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enterprise.PostNewReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/open-apis/core/enterprise/security": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "获取安保单位列表(UserAccessToken)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "信用代码",
                        "name": "serial",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "区域编码",
                        "name": "address_code",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "修改单位基本信息(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enterprise.PutSecurityReq"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "新增安保单位(UserAccessToken)",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enterprise.PostSecurityReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "enterprise.PostNewReq": {
            "type": "object",
            "properties": {
                "address_code": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "county": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "enterprise_name": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "site": {
                    "type": "string"
                },
                "staff_size": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "enterprise.PostSecurityReq": {
            "type": "object",
            "properties": {
                "address_code": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "county": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "site": {
                    "type": "string"
                },
                "staff_size": {
                    "type": "string"
                }
            }
        },
        "enterprise.PutCorporateReq": {
            "type": "object",
            "properties": {
                "corporatePk": {
                    "type": "string"
                },
                "pk": {
                    "type": "string"
                }
            }
        },
        "enterprise.PutSecurityReq": {
            "type": "object",
            "properties": {
                "address_code": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "county": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pk": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "site": {
                    "type": "string"
                },
                "staff_size": {
                    "type": "string"
                }
            }
        },
...
    }
}